Feat: types 세팅
