+++ apps/team-1/src/utils/api.ts
@@ -1,39 +1,7 @@
-import PocketBase, { RecordModel } from "pocketbase";
+import { filteredLogsTypes, userTypes } from "@/types";
+import PocketBase from "pocketbase";
 
 const pb = new PocketBase("http://54.180.1.20:8090");
-
-export interface userTypes {
-  username: string;
-  email: string;
-  emailVisibility?: boolean;
-  password: string;
-  passwordConfirm: string;
-  disable: boolean;
-  description?: string;
-  sfaclogUrl?: string;
-  category?: (
-    | "frontend"
-    | "backend"
-    | "data"
-    | "server"
-    | "dba"
-    | "logs"
-    | "android"
-  )[];
-  sns?: {
-    email?: string;
-    github?: string;
-    instagram?: string;
-    sfacfolio?: string;
-    rocketpunch?: string;
-    youtube?: string;
-  };
-}
-
-export interface loginUserTypes {
-  email: string;
-  password: string;
-}
 
 export const handleSignup = (data: userTypes) =>
   pb.collection("users").create(data);
@@ -60,14 +28,8 @@ export const handleGetAllTags = async () =>
 
 export const handleGetTargetLogs = async (
   target: string,
-): Promise<logsTypes[]> =>
+): Promise<filteredLogsTypes[]> =>
   await pb.collection("tag").getFullList({
     filter: `tagTitle='${target}'`,
     expand: "logId",
   });
-
-export interface logsTypes extends RecordModel {
-  expand: {
-    logId: { title: string };
-  };
-}
